hw1/requirements.txt
colorama==0.4.6
pillow==11.2.1
setuptools==75.8.0
tqdm==4.67.1
wheel==0.45.1
